MINOR: bump streams quickstart pom versions and add to list in gradle.properties (#13064)
The three pom files for the Streams quickstart also need to bump their versions after a branch cut. For some reason these are included (late) in the Release Process guide, but are missing from the list of what to update when bumping the version in gradle.properties. This commit adds the missing files to this list to help future RMs locate all the required version changes

Reviewers: Ismael Juma <ismael@juma.me.uk>