MINOR: fix fault handling in ControllerServer and KafkaServer (#13331)
This PR has two fixes for fault handling. One makes fault handling more
strict on the controller; the other makes it a bit less strict on the
(ZK-based) broker.

In ControllerServer, invoke a fatal fault handler in the catch block of
the ControllerServer.startup() function. This will protect us from cases
where unwinding startup would otherwise encounter a deadlock, or be too
slow.  This is the same reason why we made controller fault handlers
call halt() instead of exit() in KAFKA-14693.  In a sense, this JIRA is
a continuation of that fix since it turns some cases that would
previously have been handled by calling shutdown() into cases where we
invoke a fault handler which will call halt() directly.

In KafkaServer,  when we are in migration-from-zk mode and we create a
RaftManager, it should have a fault handler which simply calls
shutdown() on the broker rather than invoking
ProcessTerminatingFaultHandler. This fixes the bug where we could invoke
the process terminating fault handler from junit tests.

Reviewers: Ron Dagostino <rndgstn@gmail.com>