KAFKA-14834: [8/N] Propagate `isLatest` as part of `Change` (#13564)
Part of KIP-914.

This PR adds an additional boolean isLatest into Change which specifies whether the new value is the latest for its key. For un-versioned stores, isLatest is always true. For versioned stores, isLatest is true if the value has the latest timestamp seen for the key, else false. This boolean will be used by processors such as the table repartition map processor to determine when a record is out-of-order and should be dropped (when processing a versioned table).  This PR updates the table repartition map processor accordingly, and also adds test coverage for table filter.

Reviewers: Matthias J. Sax <matthias@confluent.io>